Mais validação adicionada

+++ 05 - JSP/Avaliação/Blog/src/main/webapp/rejeitarComentario.jsp
@@ -7,28 +7,30 @@
 </head>
 <body>
 <%
-    if (session.getAttribute("tipo").toString().equals("A")) {
+    if (session.getAttribute("tipo") != null) {
+        if (session.getAttribute("tipo").toString().equals("A")) {
 
-        String cdComentario = request.getParameter("cd_comentario");
+            String cdComentario = request.getParameter("cd_comentario");
 
-        // Efetuar conexão
-        Conexao c = new Conexao();
+            // Efetuar conexão
+            Conexao c = new Conexao();
 
-        // SQL
-        String sql = "DELETE FROM comentario WHERE cd_comentario = ?";
+            // SQL
+            String sql = "DELETE FROM comentario WHERE cd_comentario = ?";
 
-        // PreparedStatement
-        try (PreparedStatement pstmt = c.efetuarConexao().prepareStatement(sql)) {
+            // PreparedStatement
+            try (PreparedStatement pstmt = c.efetuarConexao().prepareStatement(sql)) {
 
-            pstmt.setInt(1, Integer.parseInt(cdComentario) );
+                pstmt.setInt(1, Integer.parseInt(cdComentario) );
 
-            pstmt.execute();
+                pstmt.execute();
 
-            // Passar os parâmetros para o SQL
-            response.sendRedirect("analiseComentarios.jsp");
+                // Passar os parâmetros para o SQL
+                response.sendRedirect("analiseComentarios.jsp");
 
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+            } catch (Exception e) {
+                System.out.println(e.getMessage());
+            }
         }
     } else {
         response.sendRedirect("index.jsp");
